backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
anthropic==0.18.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
langchain==0.1.0
langchain-anthropic==0.1.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3